# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления списком избранного пользователей",
        "title": "Favorites Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8083",
    "basePath": "/",
    "paths": {
        "/favorites": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает все варианты продуктов, добавленные в избранное авторизованным пользователем.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "favorites"
                ],
                "summary": "Получение списка избранных товаров",
                "responses": {
                    "200": {
                        "description": "favorites: array of ProductVariant",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "invalid user_id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "failed to get favorites",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/favorites/{product_variant_id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Добавляет указанный продукт (product_variant_id) в список избранного для авторизованного пользователя. Перед добавлением проверяется, что вариант продукта существует и активен.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "favorites"
                ],
                "summary": "Добавление товара в избранное",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID варианта товара",
                        "name": "product_variant_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status: added to favorites",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "invalid product_variant_id or user_id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "product variant does not exist or inactive",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "failed to add to favorites",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет указанный продукт (product_variant_id) из списка избранного для авторизованного пользователя.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "favorites"
                ],
                "summary": "Удаление товара из избранного",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID варианта товара",
                        "name": "product_variant_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status: removed from favorites",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "invalid product_variant_id or user_id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "failed to remove from favorites",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}